Added a parse script for HTML in the database and cleaned up the content table
THis script parses the HTML in the content table but does nothing to
the content except checking that the content refers to a structure that
exists in the structure table. If a structure does not exists, the
content is an orphan and is deleted. A fresh dump is also added